move from int to float32 in mainly structures, improve one TODO in code

+++ pkg/ledger/bills.go
@@ -11,8 +11,8 @@ const (
 
 // Bill type with fields
 type Bill struct {
-	Company     string `yaml:"name"`
-	Type        string `yaml:"type"`
-	Amount      int    `yaml:"amount"`
-	Description string `yaml:"description"`
+	Company     string  `yaml:"name"`
+	Type        string  `yaml:"type"`
+	Amount      float32 `yaml:"amount"`
+	Description string  `yaml:"description"`
 }
